Restructure verification components
