+++ crates/kilt-dip-primitives/src/merkle/v0/dip_subject_state/tests.rs
@@ -1,0 +1,41 @@
+// KILT Blockchain – https://botlabs.org
+// Copyright (C) 2019-2023 BOTLabs GmbH
+
+// The KILT Blockchain is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// The KILT Blockchain is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+// If you feel like getting in touch with us, you can do so at info@botlabs.org
+
+mod dip_revealed_details_and_unverified_did_signature {
+	#[test]
+	fn verify_signature_time_successful() {
+		unimplemented!()
+	}
+
+	#[test]
+	fn verify_signature_time_too_old() {
+		unimplemented!()
+	}
+}
+
+mod dip_revealed_details_and_verified_did_signature_freshness {
+	#[test]
+	fn retrieve_signing_leaf_for_payload_successful() {
+		unimplemented!()
+	}
+
+	#[test]
+	fn retrieve_signing_leaf_for_payload_no_key_present() {
+		unimplemented!()
+	}
+}
