+++ crates/kilt-dip-primitives/src/merkle/v0/relay_state/tests.rs
@@ -1,0 +1,51 @@
+// KILT Blockchain – https://botlabs.org
+// Copyright (C) 2019-2023 BOTLabs GmbH
+
+// The KILT Blockchain is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// The KILT Blockchain is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+// If you feel like getting in touch with us, you can do so at info@botlabs.org
+
+mod relay_did_dip_proof {
+	#[test]
+	fn verify_relay_header_with_block_hash_successful() {
+		unimplemented!()
+	}
+
+	#[test]
+	fn verify_relay_header_with_block_hash_wrong_hash() {
+		unimplemented!()
+	}
+
+	#[test]
+	fn verify_relay_header_with_block_hash_wrong_header() {
+		unimplemented!()
+	}
+}
+
+mod relay_dip_did_proof_with_verified_relay_state_root {
+	#[test]
+	fn verify_provider_head_proof_successful() {
+		unimplemented!()
+	}
+
+	#[test]
+	fn verify_provider_head_proof_wrong_hasher() {
+		unimplemented!()
+	}
+
+	#[test]
+	fn verify_provider_head_proof_wrong_header_type() {
+		unimplemented!()
+	}
+}
